labeler/pyproject.toml
[tool.poetry]
name = "labeler"
version = "0.1.0"
description = ""
authors = ["xetera <xetera@kiyomi.io>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
black = "^21.12b0"
mypy = "^0.930"
face-recognition = "^1.3.0"
aiohttp = "^3.8.1"
celery = {extras = ["redis"], version = "^5.2.2"}
python-dotenv = "^0.19.2"
ImageHash = "^4.2.1"
requests = "^2.26.0"
numpy = "^1.21.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
